+++ include/class-cron-list-table.php
@@ -55,6 +55,13 @@ class Cron_List_Table extends WP_List_Table {
             echo '</li>';
 
             echo '<li>';
+            echo '<strong>', __( 'Delete posts by title', 'bulk-delete' ), '</strong>', ' - ';
+            echo __( 'Adds the ability to delete posts based on title', 'bulk-delete' );
+            echo ' <a href = "http://sudarmuthu.com/wordpress/bulk-delete/pro-addons#bulk-delete-by-title">', __( 'More Info', 'bulk-delete' ), '</a>.';
+            echo ' <a href = "http://sudarmuthu.com/wordpress/bulk-delete/pro-addons#bulk-delete-by-title">', __( 'Buy now', 'bulk-delete' ), '</a>';
+            echo '</li>';
+
+            echo '<li>';
             echo '<strong>', __('Schedule auto delete of Posts by Categories', 'bulk-delete'), '</strong>', ' - ';
             echo __('Adds the ability to schedule auto delete of posts based on categories', 'bulk-delete');
             echo ' <a href = "http://sudarmuthu.com/wordpress/bulk-delete/pro-addons#bulk-delete-schedule-categories">', __('More Info', 'bulk-delete'), '</a>.';