+++ HISTORY.md
@@ -5,6 +5,7 @@
 - New: Ability to delete all sticky posts from "Post Status" module
 - Tweak: Moved all option page to a separate top menu
 - Fix: Fix undefined notices and strict warnings
+- New: Ability to delete posts by title
 
 ### 2013-12-08 - v4.3 - (Dev time: 2 hours) ###
 - New: Ability to delete custom post type posts by categories